create frame_analysis_model.dart
